V 0.1-SNAPSHOT-C
V 0.1-SNAPSHOT-C Redoes How data is stored.
It removes outside interactions of the file system by replacing the deprecated ReadData and WriteData methods with the two new data classes VaultData and PlayerData. This was done to reduce code length and allow for tweaking of internal data structures without the need for code rewires. The Data class will eventually need an update to allow multi tread utilization and asynchronous reads and writes.